Fix Pack 003: LUNATIC FORCE [Stage 5]

+++ so/shop.c
@@ -157,7 +157,7 @@ int shop_main(void)
 
     move(0, 0);
     prints("\x1b[1;33;42m%*s" BOARDNAME "    商店街%*s\x1b[m\n\n",
-           (d_cols >> 1) + 36 - sizeof(BOARDNAME), "", ((d_cols+1) >> 1) + 33, "");
+           (d_cols >> 1) + 36 - (int)(unsigned int)sizeof(BOARDNAME), "", ((d_cols+1) >> 1) + 33, "");
     prints("                     您的現金如下：\n\n                     夢幣：%d 元", money);
     move(10, 0);
